+++ main.go
@@ -25,7 +25,7 @@ import (
 )
 
 const (
-	Version   = "1.0.1"
+	Version   = "1.1.0"
 	BuildName = "noesis"
 )